KB-5130 | DEV| Assessment | BE | Enhancement in Consumption Logic for the QuestionWeightage Assessment Type
1. Assessment Bug fix.

+++ src/main/java/org/sunbird/assessment/service/AssessmentServiceV5Impl.java
@@ -821,7 +821,7 @@ public class AssessmentServiceV5Impl implements AssessmentServiceV5 {
                 }
             }
             if(practiceAssessmentFlag){
-                res.put(Constants.OVERALL_RESULT, (correct / blank+correct+inCorrect) * 100);
+                res.put(Constants.OVERALL_RESULT, ((double)correct / (double)(blank+correct+inCorrect)) * 100);
             }else{
                 res.put(Constants.OVERALL_RESULT, totalResult / sectionLevelResults.size());
             }
